# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: OrigCMS API
    description: OrigCMS is a modern Content Management System for video platforms.
    contact:
        name: GodCong
        url: https://github.com/origadmin
        email: waitforadding@gmail.com
    license:
        name: MIT
        url: https://origadmin/application/origcms/blob/master/LICENSE
    version: Version from annotation
servers:
    - url: http://localhost:18080
    - url: https://localhost:18080
paths:
...
    /api/v1/admin/stats/medias:
        get:
            tags:
                - AdminService
            operationId: AdminService_GetMediaStats
            parameters:
                - name: period
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaStatsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/channels/{channel_id}/medias:
        get:
            tags:
                - ChannelService
            operationId: ChannelService_GetChannelMedias
            parameters:
                - name: channel_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetChannelMediasResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - ChannelService
            operationId: ChannelService_AddChannelMedia
            parameters:
                - name: channel_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.AddChannelMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.AddChannelMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/channels/{channel_id}/medias/{media_id}:
        delete:
            tags:
                - ChannelService
            operationId: ChannelService_RemoveChannelMedia
            parameters:
                - name: channel_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: media_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.RemoveChannelMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias:
        get:
            tags:
                - MediaService
            description: ListMedias returns a list of medias.
            operationId: MediaService_ListMedias
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_token
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: category_id
                  in: query
                  schema:
                    type: string
                - name: user_id
                  in: query
                  schema:
                    type: string
                - name: order_by
                  in: query
                  schema:
                    type: string
                - name: descending
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.ListMediasResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - MediaService
            description: CreateMedia creates a new media.
            operationId: MediaService_CreateMedia
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.CreateMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.CreateMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias/encoding/tasks:
        get:
            tags:
                - MediaService
            description: ListAllEncodingTasks returns a flat list of all encoding tasks (admin view).
            operationId: MediaService_ListAllEncodingTasks
            parameters:
                - name: status
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: media_id
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.ListAllEncodingTasksResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias/{id}:
        get:
            tags:
                - MediaService
            description: GetMedia returns a media by ID.
            operationId: MediaService_GetMedia
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        delete:
            tags:
                - MediaService
            description: DeleteMedia deletes a media.
            operationId: MediaService_DeleteMedia
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.DeleteMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias/{id}/variants:
        get:
            tags:
                - MediaService
            description: GetMediaVariants returns the available variants for a media.
            operationId: MediaService_GetMediaVariants
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaVariantsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/medias/{id}/view:
        post:
            tags:
                - MediaService
            description: IncrementViewCount increments the view count of a media.
            operationId: MediaService_IncrementViewCount
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.IncrementViewCountRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.IncrementViewCountResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/medias/{media.id}:
        put:
            tags:
                - MediaService
            description: UpdateMedia updates an existing media.
            operationId: MediaService_UpdateMedia
            parameters:
                - name: media.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.UpdateMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.UpdateMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/playlists/{playlist_id}/medias:
        post:
            tags:
                - PlaylistService
            operationId: PlaylistService_AddPlaylistMedia
            parameters:
                - name: playlist_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.AddPlaylistMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.AddPlaylistMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/playlists/{playlist_id}/medias/{media_id}:
        delete:
            tags:
                - PlaylistService
            operationId: PlaylistService_RemovePlaylistMedia
            parameters:
                - name: playlist_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: media_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.RemovePlaylistMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
components:
    schemas:
        api.v1.services.media.AddChannelMediaRequest:
            type: object
            properties:
                channel_id:
                    type: string
                media_id:
                    type: string
...
        api.v1.services.media.GetChannelMediasResponse:
            type: object
            properties:
                total:
                    type: integer
                    format: int32
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.Media'
                page:
                    type: integer
                    format: int32
                page_size:
                    type: integer
                    format: int32
...
        api.v1.services.media.GetMediaStatsResponse:
            type: object
            properties:
                total_uploads:
                    type: integer
                    format: int32
                total_views:
                    type: integer
                    format: int32
                daily_stats:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.StatPoint'
...
        google.rpc.Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/google.protobuf.Any'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
    securitySchemes:
        Basic:
            type: http
            scheme: basic
        Bearer:
            type: apiKey
            name: Authorization
            in: header
...